# SNAP optimizations - AoSoA data layouts, chem snap indexing conventions, misc perf improvements