**Summary**

This PR provides various performance optimizations to the evaluation of the SNAP potential within the Kokkos package when using the Cuda backend (though it should generalize well to any GPU), focusing on routines for group theory contractions. It also modifies the form of data structures for chem snap with the intent of making future data layout transformations easier (described in greater detail below). This does further bifurcate the CPU/GPU code paths, which I discuss under Implementation Notes.

There is a perf regression on the CPU, appears to be because of the View rank change, which could be rolled back.

Performance notes below. For robustness, I note that `master` is specifically commit `7e2f29bb6576f68bdab06e8ffc8cc80cc6f85f19` at time of filing.

Performance improvements on V100-16GB:

|Input file  |  Timesteps/sec master | Timesteps/sec PR | Perf improvement | Comment |
| ------------- | ------------- | -------------| ----------- | -------- |
| 2J8 benchmark  | 239 | 278 | 1.16x | Benchmark problem size (2000 atoms) |
| 2J14 benchmark | 39.4 | 50.4 | 1.28x | ECP KPP problem size (2000 atoms) |
| in.snap.W.2940 | 1534 | 1580 | 1.03x | `/examples/snap/` single atom (128 atoms) |
| in.snap.WBe.PRB2019 | 1510 | 1540 | 1.02x | `/examples/snap/` multi-atom, no chem (128 atoms) |
| in.snap.InP.JCPA2020 | 422 | 476 | 1.13x | `/examples/snap/` multi-atom, yes chem (512 atoms) |

Performance regressions on 20 core Broadwell (used KOKKOS_ARCH=BDW appropriately), source appears to be changing rank of Views as CPU code is otherwise untouched. 

|Input file  |  Timesteps/sec master | Timesteps/sec PR | Perf change | Comment |
| ------------- | ------------- | -------------| ----------- | -------- |
| 2J8 benchmark  | 11.2 | 10.3 | 0.92x | Benchmark problem size (2000 atoms) |
| 2J14 benchmark | 1.34 | 1.25 | 0.93x | ECP KPP problem size (2000 atoms) |
| in.snap.W.2940 | 185 | 161 | 0.87x | `/examples/snap/` single atom (128 atoms) |
| in.snap.WBe.PRB2019 | 192 | 163 | 0.85x | `/examples/snap/` multi-atom, no chem (128 atoms) |
| in.snap.InP.JCPA2020 | 36.2 | 29.6 | 0.82x | `/examples/snap/` multi-atom, yes chem (512 atoms) |

**Related Issues**

Includes full support for chem snap (#1971, #2168).

**Author(s)**

Evan Weinberg (NVIDIA). eweinberg [at] nvidia.com preferred, evansweinberg [at] gmail.com for "long-lived".

**Licensing**

By submitting this pull request, I agree, that my contribution will be included in LAMMPS and redistributed under either the GNU General Public License version 2 (GPL v2) or the GNU Lesser General Public License version 2.1 (LGPL v2.1).

**Backward Compatibility**

No changes.

**Implementation Notes**

- Modifies `ComputeYi`, `ComputeZi`, `ComputeBi`, and the computation of beta to use an `AoSoA` data layout (as inspired by Cabana), adding new Views `[y/z/b]list_pack`. Representative example: `ylist(jju, iatom) -> ylist(iatom % 32, jju, iatom / 32)`, implemented in a way to avoid those mod/div in the first place (see `MDRangePolicy` comment)
- Adds data layout transformation routines to transform from/to data layouts for `ComputeUi` and `ComputeFusedDeidrj` as an AoSoA layout is not optimal there.
- Modifies Views that are effected by chem snap additions to have an extra View dimension as opposed to a flattened (quantum number) x (element) index. Opens an avenue for data layout experiments in the future. Representative example (truncating AoSoA): `ylist(elem1 * idxu_max + jju, iatom) -> ylist(jju, elem1, iatom)`. Source of performance regression on CPU.
- Switches as many GPU kernels as possible to use `MDRangePolicy` to better improve launch configurations and tiling, as well as eliminate modular arithmetic needed to un-flatten multidimensional indices.
- Splits the `ulisttot` and `ylist` data structures into separate real and imaginary structures for routines  (`ulisttot_[re/im]`, `ylist_pack_[re/im]`) to promote coalescing for `atomicAdd` in `ComputeUi` and `ComputeYi`, respectively.
- Optimizations to the order of instructions in `ComputeUi` and `ComputeFusedDeidrj`.
- Re-ordered the declarations and definition of routines to better separate CPU- and GPU-specific routines.

Correctness has been verified over several configurations (listed in timings above), including verifying range bounds re:conversation in #2168.

Headaches for conversation:

* We're now maximally bifurcating CPU and GPU routines (i.e., there are literally `compute_yi` vs `compute_yi_cpu`, `compute_bi` vs `compute_bi_cpu`, etc). There are tentative avenues to reconcile this, namely using an AoSoA layout on CPU and GPU, CPU inner "A" of 1 (vs 32 for GPU), and making said numbers compile-time via `constexpr` magic.
* The CPU backend has taken a perf hit, only possible source could be changing the rank of the Views. This could be reverted, so let me know your thoughts.

I haven't investigated either of these yet because @stanmoore1 has mentioned CPU-side optimizations, so I didn't want to kick any tires before getting his input.

**Post Submission Checklist**

- [ ] The feature or features in this pull request is complete
- [x] Licensing information is complete
- [ ] Corresponding author information is complete
- [ ] The source code follows the LAMMPS formatting guidelines
- [x] The feature has been verified to work with the conventional build system
- [x] The feature has been verified to work with the CMake based build system

**Further Information, Files, and Links**

None.