@weinbe2 2020-06-19
FYI @athomps @charlessievers @akkamesh 

While I don't see it as excusing the CPU perf regression, I noticed the non-Kokkos SNAP implementation sees far better performance on the CPU than either the current `master` or this PR.