@charlessievers 2020-06-19
> We're now maximally bifurcating CPU and GPU routines 
> While I don't see that as excusing the CPU perf regression, it may suggest using the non-Kokkos backend for production CPU runs:

I was on a similar page to use non-Kokkos for CPU production runs. This makes me less concerned about the bifurcation between CPU and GPU routines. The difference in architectures begs for routine bifurcation to squeeze performance from GPUs.