@charlessievers 2020-06-19
@weinbe2 Why exactly does increasing the dimensionality increase performance for GPUs?