@weinbe2 2020-06-19
> > The decrease on CPU will not be a problem for them, since the Kokkos version is already a lot slower than non-Kokkos
> 
> This is true. I'm OK with the bifurcation for now. Part of me wants to start fresh with MPI-only SNAP and make another Kokkos optimized version for OpenMP/SIMD. Then patch that along side the GPU version using the bifurcation to our advantage :) Performance portability seems to only go so far.

For what it's worth, if the SIMD approach translates to warp-level parallelism against the Cuda backend, you'll be able to start re-converging the codes!