@charlessievers 2020-06-20
@weinbe2 Thanks for the detailed explanation, I appreciate the depth. I am guessing 32 comes from the number of threads in a block. Is 32 universal across GPUs, NVIDIA GPUs, or just the GPUs we care about for ECP?