@weinbe2 2020-06-20
> @weinbe2 Thanks for the detailed explanation, I appreciate the depth. I am guessing 32 comes from the number of threads in a block. Is 32 universal across GPUs, NVIDIA GPUs, or just the GPUs we care about for ECP?

Each block can then have far more than 32 threads (I think up to 2048 on Volta), which can be configured at runtime. 32 comes from the number of threads in a warp, which is the granularity at which work is done on V100 (thus it matters for the ECP now) in the sense of the single instruction, multiple thread (SIMT) model of GPU programming, i.e., you get optimal performance if each of the threads executes the same instruction (thus the 32s... everywhere, in the PR). 32 threads/warp holds for all other NVIDIA GPUs thus far. Threads -> Warps -> Blocks corresponds to a hierarchy of locality---threads within a block can synchronize and communicate through low-latency shared memory, threads within a warp can communicate/synchronize with even-lower-latency warp instructions.

Threads are also grouped into warps on AMD GPUs (though I think they go by a different name), but iirc they have 64 threads in a warp? Not sure about their upcoming GPUs, or if it's even been announced. The AoSoA layout can be adopted to whichever granularity by modifying the 32 to whatever value is appropriate (thus the applicability to the ECP in the future)---it also generalizes to vector instructions on CPUs in principle, but there are a handful of reasons why CPU compilers may not assume they can generate vector instructions...