@stanmoore1 2020-06-22
I measured 22% speedup for 2J14. Also confirmed correct numerics.